backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
flask_pymongo
opencv-python
Werkzeug
python-dotenv
Celery
redis
